move conditional_produces to decorators and add env variable to template
