+++ app/decorators.py
@@ -10,10 +10,15 @@ import os
 
 from flask import request, make_response, current_app, abort, jsonify
 from functools import wraps
+from flask_negotiate import produces
 
 from app import is_sentry_enabled
 from app.auth import verify_token
 from app.exceptions import TokenError
+
+# content negotiation toggler
+ENFORCE_CONTENT_NEGOTIATION = os.getenv("ENFORCE_CONTENT_NEGOTIATION", "True").lower() == "true"
+
 
 if is_sentry_enabled():
     import sentry_sdk
@@ -108,3 +113,11 @@ def keycloak_protected(f):
         return f(*args, **kwargs)
 
     return decorated_function
+
+
+def conditional_produces(mime_type):
+    def decorator(func):
+        if ENFORCE_CONTENT_NEGOTIATION:
+            return produces(mime_type)(func)
+        return func
+    return decorator
