+++ env.template
@@ -18,3 +18,4 @@ KEYCLOAK_TOKEN_URI="https://example.org/realms/my-auth/protocol/openid-connect/t
 KEYCLOAK_INTROSPECT_URI="https://example.org/realms/my-auth/protocol/openid-connect/token/introspect"
 SUPPORTED_API_VERSION="1.0"
 OPEN_API_SERVER_URL=https://your-custom-url.com
+ENFORCE_CONTENT_NEGOTIATION=False
